# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MaiEcho 是一个用于音乐评论和分析的服务器应用程序。",
        "title": "MaiEcho API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/analysis/batch": {
            "post": {
                "description": "触发针对多个歌曲ID(GameID)的LLM分析流程",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "批量分析歌曲",
                "parameters": [
                    {
                        "description": "Batch Analysis Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller.BatchAnalysisRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/analysis/songs/{id}": {
            "get": {
                "description": "获取指定歌曲(GameID)的最新分析结果",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "获取分析结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_xumoe-c_maiecho_server_internal_model.AnalysisResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "触发针对指定歌曲ID(GameID)的LLM分析流程",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analysis"
                ],
                "summary": "分析歌曲",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/collect": {
            "post": {
                "description": "针对特定关键词或GameID启动数据收集任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collector"
                ],
                "summary": "触发数据收集任务",
                "parameters": [
                    {
                        "description": "Collection Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller.CollectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/collect/backfill": {
            "post": {
                "description": "用于初始化数据源，为无数据的歌曲收集评论数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collector"
                ],
                "summary": "触发回填数据收集",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/songs": {
            "get": {
                "description": "支持分页和多种筛选条件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "获取歌曲列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "版本",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "类型 (DX/Standard)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "流派",
                        "name": "genre",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "最小定数",
                        "name": "min_ds",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "最大定数",
                        "name": "max_ds",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否新歌",
                        "name": "is_new",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_xumoe-c_maiecho_server_internal_model.SongListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "向数据库添加新歌曲",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "创建歌曲",
                "parameters": [
                    {
                        "description": "Song JSON",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_xumoe-c_maiecho_server_internal_model.Song"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_xumoe-c_maiecho_server_internal_model.Song"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/songs/aliases/refresh": {
            "post": {
                "description": "从YuzuChan API获取最新的歌曲别名",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "刷新歌曲别名",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/songs/sync": {
            "post": {
                "description": "从Diving-Fish API获取并更新歌曲数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "同步歌曲数据",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/songs/{id}": {
            "get": {
                "description": "获取歌曲及其谱面详情 (通过 GameID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "获取歌曲详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_xumoe-c_maiecho_server_internal_model.Song"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/system/status": {
            "get": {
                "description": "获取服务器运行状态、资源使用情况等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取系统状态",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_xumoe-c_maiecho_server_internal_status.SystemStatus"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_xumoe-c_maiecho_server_internal_model.AnalysisResult": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "difficulty_analysis": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating_advice": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "target_id": {
                    "type": "integer"
                },
                "target_type": {
                    "description": "Song or Chart",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_xumoe-c_maiecho_server_internal_model.Chart": {
            "type": "object",
            "properties": {
                "avg_achievement": {
                    "type": "number"
                },
                "avg_dx": {
                    "type": "number"
                },
                "charter": {
                    "description": "NotesDesigner",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "difficulty": {
                    "description": "Basic, Advanced, Expert, Master, Re:Master",
                    "type": "string"
                },
                "ds": {
                    "description": "Internal decimal level",
                    "type": "number"
                },
                "fit_diff": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "description": "13, 13+, 14, etc.",
                    "type": "string"
                },
                "notes": {
                    "description": "JSON array: [tap, hold, slide, touch, break]",
                    "type": "string"
                },
                "sample_count": {
                    "type": "integer"
                },
                "song_id": {
                    "type": "integer"
                },
                "std_dev": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_xumoe-c_maiecho_server_internal_model.Song": {
            "type": "object",
            "properties": {
                "aliases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_xumoe-c_maiecho_server_internal_model.SongAlias"
                    }
                },
                "artist": {
                    "type": "string"
                },
                "bpm": {
                    "type": "number"
                },
                "charts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_xumoe-c_maiecho_server_internal_model.Chart"
                    }
                },
                "cover_url": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "genre": {
                    "type": "string"
                },
                "id": {
                    "description": "来自 Diving-Fish 的 ID",
                    "type": "integer"
                },
                "is_new": {
                    "type": "boolean"
                },
                "release_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "description": "DX or 标准",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "version": {
                    "description": "乐曲更新版本",
                    "type": "string"
                }
            }
        },
        "github_com_xumoe-c_maiecho_server_internal_model.SongAlias": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "song_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_xumoe-c_maiecho_server_internal_model.SongListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_xumoe-c_maiecho_server_internal_model.Song"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_xumoe-c_maiecho_server_internal_status.SystemStatus": {
            "type": "object",
            "properties": {
                "active_tasks": {
                    "type": "integer"
                },
                "goroutines": {
                    "type": "integer"
                },
                "last_log_entry": {
                    "type": "string"
                },
                "memory_usage_mb": {
                    "type": "integer"
                },
                "uptime": {
                    "$ref": "#/definitions/time.Duration"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "internal_controller.BatchAnalysisRequest": {
            "type": "object",
            "required": [
                "game_ids"
            ],
            "properties": {
                "game_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "internal_controller.CollectRequest": {
            "type": "object",
            "properties": {
                "game_id": {
                    "type": "integer"
                },
                "keyword": {
                    "type": "string"
                }
            }
        },
        "time.Duration": {
            "type": "integer",
            "format": "int64",
            "enum": [
                -9223372036854775808,
                9223372036854775807,
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000
            ],
            "x-enum-varnames": [
                "minDuration",
                "maxDuration",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour"
            ]
        }
    }
}